% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vir.R
\docType{data}
\name{grey128}
\alias{grey128}
\title{128 grey palette - 128x3 integer matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 128 rows and 3 columns.
}
\usage{
grey128
}
\description{
128 grey palette - 128x3 integer matrix
}
\keyword{datasets}
