% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_gif_core}
\alias{write_gif_core}
\title{Write a numeric matrix or array to a GIF file}
\usage{
write_gif_core(vec, dims, filename, convert_to_row_major = TRUE,
  flipy = FALSE, invert = FALSE, intensity_factor = 1, pal = NULL)
}
\arguments{
\item{vec}{numeric 2d matrix}

\item{dims}{integer vector of length 2 i.e. \code{c(nrow, ncol)} for a matrix/grey image}

\item{filename}{output filename e.g. "example.ppm"}

\item{convert_to_row_major}{Convert to row-major order before output. R stores matrix
and array data in column-major order. In order to output row-major order (as
expected by PGM/PPM image format) data ordering must be converted. If this argument
is set to FALSE, then image output will be faster (due to fewer data-ordering operations, and
better cache coherency) but the image will be transposed. Default: TRUE}

\item{flipy}{By default, the position [0, 0] is considered the top-left corner of the output image.
Set flipy = TRUE for [0, 0] to represent the bottom-left corner.  This operation
is very fast and has negligible impact on overall write speed.
Default: flipy = FALSE.}

\item{invert}{invert all the pixel brightness values - as if the image were
converted into a negative. Dark areas become bright and bright areas become dark.
Default: FALSE}

\item{intensity_factor}{Multiplication factor applied to all values in image
(note: no checking is performed to ensure values remain in range [0, 1]).
If intensity_factor <= 0, then automatically determine (and apply) a multiplication factor
to set the maximum value to 1.0. Default: intensity_factor = 1.0}

\item{pal}{integer matrix of size 128x3 or 256x3 with values in the range [0, 255]. Each
row represents the r, g, b colour for a given grey index value. This
GIF writer only supports 128 colours, so a 256x3 palette is reduced
to 128 colours by selecting every second colour. If supplied with a
128-colour-palette then it is used as-is.}
}
\description{
Write a numeric matrix or array to a GIF file
}
\details{
Write a numeric matrix to a GIF file as fast as I can - meaning
that corners are cut to make it happen quickly:

\itemize{
\item{Data is not compressed.}
}
}
