% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pnm.R
\name{write_pnm}
\alias{write_pnm}
\title{Write a numeric matrix or array to a NETPBM PNM file}
\usage{
write_pnm(
  data,
  filename,
  convert_to_row_major = TRUE,
  flipy = FALSE,
  invert = FALSE,
  intensity_factor = 1,
  pal = NULL
)
}
\arguments{
\item{data}{numeric 2d matrix or 3d array (with 3 planes)}

\item{filename}{output filename e.g. "example.ppm"}

\item{convert_to_row_major}{Convert to row-major order before output. R stores matrix
and array data in column-major order. In order to output row-major order (as
expected by PGM/PPM image format) data ordering must be converted. If this argument
is set to FALSE, then image output will be faster (due to fewer data-ordering operations, and
better cache coherency) but the image will be transposed. Default: TRUE}

\item{flipy}{By default, the position [0, 0] is considered the top-left corner of the output image.
Set flipy = TRUE for [0, 0] to represent the bottom-left corner.  This operation
is very fast and has negligible impact on overall write speed.
Default: flipy = FALSE.}

\item{invert}{invert all the pixel brightness values - as if the image were
converted into a negative. Dark areas become bright and bright areas become dark.
Default: FALSE}

\item{intensity_factor}{Multiplication factor applied to all values in image
(note: no checking is performed to ensure values remain in range [0, 1]).
If intensity_factor <= 0, then automatically determine (and apply) a multiplication factor
to set the maximum value to 1.0. Default: intensity_factor = 1.0}

\item{pal}{integer matrix of size 256x3 with values in the range [0, 255]. Each
row represents the r, g, b colour for a given grey index value. Only used
if \code{data} is a matrix}
}
\description{
Write a numeric matrix or array to a NETPBM PNM file
}
