% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_pnm_core}
\alias{write_pnm_core}
\title{Write a vector of numeric data to a PNM file}
\usage{
write_pnm_core(vec, dims, filename, convert_to_row_major = TRUE,
  flipy = FALSE, invert = FALSE, intensity_factor = 1, pal = NULL)
}
\arguments{
\item{vec}{numeric vector of data}

\item{dims}{integer vector of length 2 i.e. \code{c(nrow, ncol)} for a matrix/grey image and of
length 3 i.e \code{c(nrow, ncol, 3)} for array/RGB output.}

\item{filename}{output filename e.g "example.pgm"}

\item{convert_to_row_major}{Convert to row-major order before output. R stores matrix
and array data in column-major order. In order to output row-major order (as
expected by most image formats) data ordering must be converted. If this argument
is set to FALSE, then image output will be faster (due to fewer data-ordering operations, and
better cache coherency) but the image will appear transposed. Default: TRUE}

\item{flipy}{By default, the position [0, 0] is considered the top-left corner of the output image.
Set flipy = TRUE for [0, 0] to represent the bottom-left corner.  This operation
is very fast and has negligible impact on overall write speed.
Default: flipy = FALSE.}

\item{invert}{invert all the pixel brightness values - as if the image were
converted into a negative. Dark areas become bright and bright areas become dark.
Default: FALSE}

\item{intensity_factor}{Multiplication factor applied to all values in image
(note: no checking is performed to ensure values remain in range [0, 1]).
If intensity_factor <= 0, then automatically determine (and apply) a multiplication factor
to set the maximum value to 1.0. Default: intensity_factor = 1.0}

\item{pal}{integer matrix of size 256x3 with values in the range [0, 255]. Each
row represents the r, g, b colour for a given grey index value. Only used
if \code{vec} is a matrix}
}
\description{
Write a vector of numeric data to a PNM file
}
