% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm_wrapper.R
\name{write_ppm}
\alias{write_ppm}
\title{Write a numeric array to a PPM file}
\usage{
write_ppm(arr, filename, convert_to_row_major = TRUE,
  intensity_factor = 1)
}
\arguments{
\item{arr}{numeric array with 3 planes.}

\item{filename}{output filename e.g. "example.ppm"}

\item{convert_to_row_major}{Convert to row-major order before output. R stores matrix
and array data in column-major order. In order to output row-major order (as
expected by PGM/PPM image format) data ordering must be converted. If this argument
is set to FALSE, then image output will be faster (due to fewer data-ordering operations, and
better cache coherency) but the image will be transposed. Default: TRUE}

\item{intensity_factor}{Multiplication factor applied to all values in image
(note: no checking is performed to ensure values remain in range [0, 1]).
If intensity_factor <= 0, then automatically determine (and apply) a multiplication factor
to set the maximum value to 1.0. Default: intensity_factor = 1.0}
}
\description{
Write a numeric array to a PPM file
}
