% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{write_ppm_core}
\alias{write_ppm_core}
\title{Write a numeric array to a PPM file}
\usage{
write_ppm_core(arr, dims, filename, convert_to_row_major = TRUE,
  intensity_factor = 1)
}
\arguments{
\item{arr}{numeric array with 3 planes. Because Rcpp doesn't explicitly
handle arrays, they are treated as \code{NumericVector} and
dimensions must be handled manually.}

\item{dims}{integer vector of length 3. image dimensions: c(nrow, ncol, depth). Depth must be 3.}

\item{filename}{output filename e.g. "example.ppm"}

\item{convert_to_row_major}{Convert to row-major order before output. R stores matrix
and array data in column-major order. In order to output row-major order (as
expected by PGM/PPM image format) data ordering must be converted. If this argument
is set to FALSE, then image output will be faster (due to fewer data-ordering operations, and
better cache coherency) but the image will be transposed. Default: TRUE}

\item{intensity_factor}{Multiplication factor applied to all values in image
(note: no checking is performed to ensure values remain in range [0, 1]).
If intensity_factor <= 0, then automatically determine (and apply) a multiplication factor
to set the maximum value to 1.0. Default: intensity_factor = 1.0}
}
\description{
Write a numeric array to a PPM file
}
